doc: avoid MOLECULE_EPHEMERAL_DIRECTORY with tox
The `MOLECULE_EPHEMERAL_DIRECTORY` is used to compute
the `ephemeral_directory` variable[1]. If the environment variable
exists, molecule prefix it with the temp directory.

In config, `ephemeral_directory` is used to overwrite
`MOLECULE_EPHEMERAL_DIRECTORY`[2].

As a result:
- `MOLECULE_EPHEMERAL_DIRECTORY` is different
- it can content a /, which is not a good fit in a docker or VM name

This commit promotes a new environment variables called `TOX_ENVNAME` to
avoid any conflict.

[1]: https://github.com/ansible/molecule/blob/master/molecule/scenario.py#L182-L186
[2]: https://github.com/ansible/molecule/blob/master/molecule/config.py#L193